+++ rpc-client/src/nonblocking/rpc_client.rs
@@ -2267,6 +2267,7 @@ impl RpcClient {
     /// # Examples
     ///
     /// ```
+    /// # #![allow(deprecated)]
     /// # use solana_rpc_client_api::{
     /// #     client_error::Error,
     /// #     config::RpcLargestAccountsConfig,
@@ -2289,6 +2290,12 @@ impl RpcClient {
     /// # })?;
     /// # Ok::<(), Error>(())
     /// ```
+    #[allow(deprecated)]
+    #[deprecated(
+        since = "4.1.0",
+        note = "This uses the deprecated `getLargestAccounts` RPC method and may be removed in a \
+                future major release"
+    )]
     pub async fn get_largest_accounts_with_config(
         &self,
         config: RpcLargestAccountsConfig,